+++ examples/Infinite.hs
@@ -6,7 +6,7 @@ import Text.Earley
 grammar :: Grammar r (Prod r () Char [Maybe Char])
 grammar = mdo
   as <- rule $ pure []
-            <|> (:) <$> optional (symbol 'a') <*> as
+            <|> (:) <$> optional (token 'a') <*> as
   return as
 
 -- This grammar has an infinite number of results. We can still recognise the