+++ tests/ReversedWords.hs
@@ -6,7 +6,7 @@ import Test.Tasty.HUnit as HU
 import Text.Earley
 
 someWords :: Grammar r (Prod r () Char [String])
-someWords = return $ flip (:) <$> (map reverse <$> some (word "word")) <*> word "stop"
+someWords = return $ flip (:) <$> (map reverse <$> some (list "word")) <*> list "stop"
 
 tests :: TestTree
 tests = testGroup "Unit Tests"